#75 put asserts back to main test file as the error messages don't show the values if function is in different file
